requirements.txt
boto3==1.28.57
google-cloud-storage==2.11.0
pytest==7.4.2